Rough idea of structure, python, pil, pyplot experiment
As I am unable to get ffmpeg to pickup any DRM KMS related format
modifiers, dug into python and its modules to try and explore the
available image data and inturn infer how the data may be spreadout.

Found that there seems to be a pattern every 128 pixels. So shaped
the full image to be a very long 128 wide image.

With this have also noticed that 8 adjacent rows, each of size 128
pixels is grouped together.