A column in the life of the screen
Got one full column (of width 128 pixels) of the screen to be properly
rearranged from the spreadout screen catpure.

Also as will be visible from the code, I had converted the color
image capture into a simple mono chromatic image, to make it easy
to manipulate the image data to try and find the pattern, previously.

The spread out pattern of the pixel data in the screens buffer is
clear now.