Hurray ;-) Untiled full screen wrt ffmpeg kmsgrab
Logic updated to untile the tiled raw hwdownload based full screen
capture using ffmpeg kmsgrab.

Here the tiled storage was keeping 128x8 pixel tiles stored as a
linear memory data in row major (if I remember the convention correctly
i.e) format. IE 128 adjacent pixels are stored next to one another
linearly, then it comes down to next horizontal line and stores its 128
pixels, then comes down to next horizontal line and store its 128 pixels
like this for 8 horizontal lines. After that it goes back to first row
and repeat for the next 128x8 pixel tile and so. Onces the last tile
in the 1st row of tiles is done. Then start with 1st tile of 2nd row of
tiles, and so on.

NOTE: Chances are it may warry from one gpu family to another or from
one vendor to another. This is for a intel i7-7500's gpu.

NOTE: I am bad with names so also conventional terminologies, so term
used could be conveying a different meaning from what it actually means
in conventional lingo.